requirements.txt
aenum==3.1.16
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aioice==0.10.2
aiortc==1.14.0
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
av==16.0.1
cachetools==6.2.2
cartesia==2.0.17
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
coloredlogs==15.0.1
cryptography==46.0.3
daily-python==0.22.0
dataclasses-json==0.6.7
deepgram-sdk==4.7.0
deprecation==2.1.0
distro==1.9.0
dnspython==2.8.0
docstring_parser==0.17.0
email-validator==2.3.0
fastapi==0.122.0
fastapi-cli==0.0.16
fastapi-cloud-cli==0.5.2
fastar==0.8.0
flatbuffers==25.9.23
frozenlist==1.8.0
future==1.0.0
google-api-core==2.28.1
google-auth==2.43.0
google-cloud-speech==2.34.0
google-cloud-texttospeech==2.33.0
google-crc32c==1.7.1
google-genai==1.52.0
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.0
humanfriendly==10.0
idna==3.11
ifaddr==0.2.0
iterators==0.2.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.2
llvmlite==0.44.0
loguru==0.7.3
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.1
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.0
mypy_extensions==1.1.0
nltk==3.9.2
numba==0.61.2
numpy==2.2.6
onnxruntime==1.23.2
openai==2.8.1
opencv-python==4.12.0.88
orjson==3.11.4
packaging==25.0
pillow==11.3.0
pipecat-ai[websocket]==0.0.96
-e git+https://github.com/pipecat-ai/pipecat-flows.git@92dedda66278db4a42602e2ddfbd542d41cc7abb#egg=pipecat_ai_flows
pipecat-ai-small-webrtc-prebuilt==1.0.0
propcache==0.4.1
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic-extra-types==2.10.6
pydantic-settings==2.12.0
pydantic_core==2.41.5
pydub==0.25.1
pyee==13.0.0
Pygments==2.19.2
pylibsrtp==1.0.0
pyloudnorm==0.1.1
pyOpenSSL==25.3.0
python-dotenv==1.2.1
python-multipart==0.0.20
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
resampy==0.4.3
rich==14.2.0
rich-toolkit==0.17.0
rignore==0.7.6
rsa==4.9.1
sarvamai==0.1.21
scipy==1.16.3
sentry-sdk==2.46.0
shellingham==1.5.4
sniffio==1.3.1
soxr==0.5.0.post1
starlette==0.50.0
sympy==1.14.0
tenacity==9.1.2
tqdm==4.67.1
typer==0.20.0
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
ujson==5.11.0
urllib3==2.5.0
uvicorn==0.38.0
uvloop==0.22.1
wait_for2==0.4.1
watchfiles==1.1.1
websockets==15.0.1
yarl==1.22.0
